[DownloadsWidget] Extract file searcher into own file
